sreweekly/requirements.txt
beautifulsoup4==4.8.0
pika==1.1.0
slackclient==2.1.0
